mem-ruby: Reduce handshaking between CorePair and dir
Currently when data is downgraded by MOESI_AMD_Base-CorePair (e.g. due to a replacement)
this requires a 4-way handshake between the CorePair and the dir.
Specifically, the CorePair send a message telling the dir it'd like to downgrade then,
the dir sends an ACK back and then, the CorePair writes the data back, and finally,
the dir ACKs the writeback.
This is very inefficient and not representative of how modern protocols downgrade a request.
Accordingly, this commits updates the downgrade support such that the CorePair writes back
the data immediately and then the dir ACKs it.
Thus, this approach requires only a 2-way handshake.

Change-Id: I7ebc85bb03e8ce46a8847e3240fc170120e9fcd6